+++ src/ast/domain.ts
@@ -1,10 +1,16 @@
+import { NodePath } from "@babel/traverse";
 import * as t from "@babel/types";
+
+import { last } from "../array-helpers";
 
 export * from "@babel/types";
 export {
   isArrayExpressionElement,
   areAllObjectProperties,
   isUndefinedLiteral,
+  isGuardClause,
+  isGuardConsequentBlock,
+  isNonEmptyReturn,
   isTemplateExpression,
   templateElement,
   Primitive
@@ -27,6 +33,23 @@ function isUndefinedLiteral(
   opts?: object | null
 ): node is t.Identifier {
   return t.isIdentifier(node, opts) && node.name === "undefined";
+}
+
+function isGuardClause(path: NodePath<t.IfStatement>) {
+  const { consequent } = path.node;
+  return t.isReturnStatement(consequent) || isGuardConsequentBlock(consequent);
+}
+
+function isGuardConsequentBlock(
+  consequent: t.IfStatement["consequent"]
+): consequent is t.BlockStatement {
+  return (
+    t.isBlockStatement(consequent) && t.isReturnStatement(last(consequent.body))
+  );
+}
+
+function isNonEmptyReturn(node: t.Node) {
+  return t.isReturnStatement(node) && node.argument !== null;
 }
 
 function isTemplateExpression(node: t.Node): node is TemplateExpression {