+++ src/refactorings/flip-if-else/flip-if-else.test.ts
@@ -128,6 +128,88 @@ describe("Flip If/Else", () => {
 } else {
   doAnotherThing();
 }`
+      },
+      {
+        description: "guard clause",
+        code: `if (!isValid) return;
+
+doSomething();
+doSomethingElse();`,
+        expected: `if (isValid) {
+  doSomething();
+  doSomethingElse();
+}`
+      },
+      {
+        description: "guard clause with block statement",
+        code: `if (!isValid) {
+  return;
+}
+
+doSomething();
+doSomethingElse();`,
+        expected: `if (isValid) {
+  doSomething();
+  doSomethingElse();
+}`
+      },
+      {
+        description: "guard clause with other statements in block",
+        code: `if (!isValid) {
+  console.log("Hello");
+  return;
+}
+
+doSomething();
+doSomethingElse();`,
+        expected: `if (isValid) {
+  doSomething();
+  doSomethingElse();
+} else {
+  console.log("Hello");
+}`
+      },
+      {
+        description: "guard clause with other statements above",
+        code: `console.log("Hello");
+if (!isValid) return;
+
+doSomething();
+doSomethingElse();`,
+        selection: Selection.cursorAt(1, 16),
+        expected: `console.log("Hello");
+if (isValid) {
+  doSomething();
+  doSomethingElse();
+}`
+      },
+      {
+        description: "guard clause with returned value",
+        code: `if (!isValid) return null;
+
+doSomething();
+doSomethingElse();`,
+        expected: `if (isValid) {
+  doSomething();
+  doSomethingElse();
+} else {
+  return null;
+}`
+      },
+      {
+        description: "guard clause with returned value in block",
+        code: `if (!isValid) {
+  return null;
+}
+
+doSomething();
+doSomethingElse();`,
+        expected: `if (isValid) {
+  doSomething();
+  doSomethingElse();
+} else {
+  return null;
+}`
       }
     ],
     async ({ code, selection = Selection.cursorAt(0, 0), expected }) => {