+++ src/refactorings/flip-if-else/flip-if-else.ts
@@ -1,6 +1,7 @@
 import { Editor, Code, ErrorReason } from "../../editor/editor";
 import { Selection } from "../../editor/selection";
 import * as ast from "../../ast";
+import { last, allButLast } from "../../array-helpers";
 
 import { getNegatedBinaryOperator } from "../negate-expression/negate-expression";
 
@@ -37,15 +38,58 @@ function updateCode(code: Code, selection: Selection): ast.Transformed {
       // if a child would match the selection closer.
       if (hasChildWhichMatchesSelection(path, selection)) return;
 
-      const ifBranch = node.consequent;
-      const elseBranch = node.alternate || ast.blockStatement([]);
-      node.consequent = ast.isIfStatement(elseBranch)
-        ? ast.blockStatement([elseBranch])
-        : elseBranch;
-      node.alternate = ifBranch;
+      if (ast.isGuardClause(path)) {
+        flipGuardClause(path);
+      } else {
+        flipIfStatement(path);
+      }
+
       node.test = getNegatedIfTest(node.test);
     }
   });
+}
+
+function flipIfStatement(path: ast.NodePath<ast.IfStatement>) {
+  const ifBranch = path.node.consequent;
+  const elseBranch = path.node.alternate || ast.blockStatement([]);
+
+  path.node.consequent = ast.isIfStatement(elseBranch)
+    ? ast.blockStatement([elseBranch])
+    : elseBranch;
+  path.node.alternate = ifBranch;
+}
+
+function flipGuardClause(path: ast.NodePath<ast.IfStatement>) {
+  const ifBranch = path.node.consequent;
+  const pathsBelow = path
+    .getAllNextSiblings()
+    .filter(
+      (path): path is ast.NodePath<ast.Statement> => ast.isStatement(path)
+    );
+  const nodesBelow: ast.Statement[] = pathsBelow.map(path => path.node);
+
+  path.node.consequent = ast.blockStatement(nodesBelow);
+  path.node.alternate = flipToGuardAlternate(ifBranch);
+  pathsBelow.forEach(path => path.remove());
+}
+
+function flipToGuardAlternate(
+  consequent: ast.Statement
+): ast.BlockStatement | null {
+  if (ast.isNonEmptyReturn(consequent)) {
+    return ast.blockStatement([consequent]);
+  }
+
+  if (!ast.isGuardConsequentBlock(consequent)) return null;
+
+  const finalReturnStatement = last(consequent.body);
+  const alternateBody = ast.isNonEmptyReturn(finalReturnStatement)
+    ? consequent.body
+    : allButLast(consequent.body);
+
+  if (alternateBody.length === 0) return null;
+
+  return ast.blockStatement(alternateBody);
 }
 
 function hasChildWhichMatchesSelection(