+++ CHANGELOG.md
@@ -51,7 +51,9 @@ function doSomething(someData) {
 ### Fixed
 
 - Inline Function now says it can't inline function with many statements to assigned call expressions
-- Inline Function now works on every call expression assigned to a variable (e.g. it works on ternaries)
+- Inline Function now works on every call expression that is:
+  - assigned to a variable (e.g. `const result = isValid ? inlineMe() : "default";`)
+  - inside another call expression (e.g. `console.log(inlineMe())`)
 - Extract Variable on JSX Elements now triggers symbol rename as expected
 - Extract Variable now works on `JSXText`s