+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2019-08-18
+
 ### Changed
 
 - Flip If/Else now works better on [guard clause][guard-clause] patterns
@@ -110,7 +112,8 @@ function doSomething(someData) {
 - **[New refactoring]** Extract Variable
 - **[New refactoring]** Rename Symbol
 
-[unreleased]: https://github.com/nicoespeon/abracadabra/compare/0.3.0...HEAD
+[unreleased]: https://github.com/nicoespeon/abracadabra/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/nicoespeon/abracadabra/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/nicoespeon/abracadabra/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/nicoespeon/abracadabra/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/nicoespeon/abracadabra/compare/0.0.1...0.1.0